+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Types of changes: Added, Changed, Deprecated, Removed, Fixed, Security
 
+## [Unreleased]
+
+### Security
+
+- Add Origin/Referer validation for state-changing HTTP requests to provide additional CSRF protection
+
 ## [0.9.0] - 2026-03-14
 
 ### Added